+++ PersistencePlugin/src/com/elmakers/mine/bukkit/plugins/persistence/stores/PersistenceStore.java
@@ -13,6 +13,8 @@ public abstract class PersistenceStore
 	
 	public abstract void validateTable(PersistedClass persisted);
 	
+	public abstract void reset(PersistedClass persisted);
+	
 	public abstract boolean loadAll(PersistedClass persisted);
 	public abstract boolean save(PersistedClass persisted, Object o);